move and rename LoggerFormatter class
