+++ src/middlewared/middlewared/logger.py
@@ -3,6 +3,8 @@ from logging.config import dictConfig
 import logging.handlers
 import os
 
+from .logging.console_formatter import ConsoleLogFormatter
+
 # markdown debug is also considered useless
 logging.getLogger('MARKDOWN').setLevel(logging.INFO)
 # asyncio runs in debug mode but we do not need INFO/DEBUG
@@ -47,52 +49,6 @@ def trace(self, message, *args, **kws):
 
 logging.addLevelName(logging.TRACE, "TRACE")
 logging.Logger.trace = trace
-
-
-class LoggerFormatter(logging.Formatter):
-    """Format the console log messages"""
-
-    CONSOLE_COLOR_FORMATTER = {
-        'YELLOW': '\033[1;33m',  # (warning)
-        'GREEN': '\033[1;32m',  # (info)
-        'RED': '\033[1;31m',  # (error)
-        'HIGHRED': '\033[1;41m',  # (critical)
-        'RESET': '\033[1;m',  # Reset
-    }
-    LOGGING_LEVEL = {
-        'CRITICAL': 50,
-        'ERROR': 40,
-        'WARNING': 30,
-        'INFO': 20,
-        'DEBUG': 10,
-        'NOTSET': 0
-    }
-
-    def format(self, record):
-        """Set the color based on the log level.
-
-            Returns:
-                logging.Formatter class.
-        """
-
-        if record.levelno == self.LOGGING_LEVEL['CRITICAL']:
-            color_start = self.CONSOLE_COLOR_FORMATTER['HIGHRED']
-        elif record.levelno == self.LOGGING_LEVEL['ERROR']:
-            color_start = self.CONSOLE_COLOR_FORMATTER['HIGHRED']
-        elif record.levelno == self.LOGGING_LEVEL['WARNING']:
-            color_start = self.CONSOLE_COLOR_FORMATTER['RED']
-        elif record.levelno == self.LOGGING_LEVEL['INFO']:
-            color_start = self.CONSOLE_COLOR_FORMATTER['GREEN']
-        elif record.levelno == self.LOGGING_LEVEL['DEBUG']:
-            color_start = self.CONSOLE_COLOR_FORMATTER['YELLOW']
-        else:
-            color_start = self.CONSOLE_COLOR_FORMATTER['RESET']
-
-        color_reset = self.CONSOLE_COLOR_FORMATTER['RESET']
-
-        record.levelname = color_start + record.levelname + color_reset
-
-        return logging.Formatter.format(self, record)
 
 
 class Logger(object):
@@ -180,7 +136,7 @@ class Logger(object):
             console_handler = logging.StreamHandler()
             logging.root.setLevel(getattr(logging, self.debug_level))
             time_format = "%Y/%m/%d %H:%M:%S"
-            console_handler.setFormatter(LoggerFormatter(self.log_format, datefmt=time_format))
+            console_handler.setFormatter(ConsoleLogFormatter(self.log_format, datefmt=time_format))
             logging.root.addHandler(console_handler)
         else:
             dictConfig(self.DEFAULT_LOGGING)
