+++ src/middlewared/middlewared/logging/console_formatter.py
@@ -1,0 +1,47 @@
+import logging
+
+
+class LoggerFormatter(logging.Formatter):
+    """Format the console log messages"""
+
+    CONSOLE_COLOR_FORMATTER = {
+        'YELLOW': '\033[1;33m',  # (warning)
+        'GREEN': '\033[1;32m',  # (info)
+        'RED': '\033[1;31m',  # (error)
+        'HIGHRED': '\033[1;41m',  # (critical)
+        'RESET': '\033[1;m',  # Reset
+    }
+    LOGGING_LEVEL = {
+        'CRITICAL': 50,
+        'ERROR': 40,
+        'WARNING': 30,
+        'INFO': 20,
+        'DEBUG': 10,
+        'NOTSET': 0
+    }
+
+    def format(self, record):
+        """Set the color based on the log level.
+
+            Returns:
+                logging.Formatter class.
+        """
+
+        if record.levelno == self.LOGGING_LEVEL['CRITICAL']:
+            color_start = self.CONSOLE_COLOR_FORMATTER['HIGHRED']
+        elif record.levelno == self.LOGGING_LEVEL['ERROR']:
+            color_start = self.CONSOLE_COLOR_FORMATTER['HIGHRED']
+        elif record.levelno == self.LOGGING_LEVEL['WARNING']:
+            color_start = self.CONSOLE_COLOR_FORMATTER['RED']
+        elif record.levelno == self.LOGGING_LEVEL['INFO']:
+            color_start = self.CONSOLE_COLOR_FORMATTER['GREEN']
+        elif record.levelno == self.LOGGING_LEVEL['DEBUG']:
+            color_start = self.CONSOLE_COLOR_FORMATTER['YELLOW']
+        else:
+            color_start = self.CONSOLE_COLOR_FORMATTER['RESET']
+
+        color_reset = self.CONSOLE_COLOR_FORMATTER['RESET']
+
+        record.levelname = color_start + record.levelname + color_reset
+
+        return logging.Formatter.format(self, record)
